# api/openapi.yaml
openapi: 3.1.0
info:
  title: Messenger API
  version: 2.0.0-alpha.5
  summary: REST contract of the Messenger backend (v2).
  description: |
    Contract for the Go backend. The WebSocket protocol is described in `websocket/` and
    `websocket.md`; the entity schemas below are shared by both.

    ## Conventions

    - Base path `/api/v2`. Breaking changes need a new MAJOR version with a new base path
      (see `docs/api-compatibility.md`).
    - All IDs are **strings** in JSON. Message IDs are 64-bit Snowflake IDs, larger than
      JavaScript's safe integer limit (2^53), so they must never be parsed as numbers.
    - Timestamps are RFC 3339 UTC (`2026-10-04T12:00:00Z`).
    - Errors are `application/problem+json` (RFC 9457) with a stable `code`; clients switch on
      `code`, never on `title` or `detail`.
    - Lists use cursor pagination: `limit` plus `before`, `after` or `around`, and responses
      return `next_cursor` (and `prev_cursor` for messages). Cursors are opaque to clients
      except for message history, where they are message IDs.
    - Enums can grow in a MINOR version; clients must handle unknown values with a fallback.
    - Clients send `X-Client-Version` (app version) and `X-Client-Api-Version` (the contract
      version they were built with). A different MAJOR or a version below
      `min_client_api_version` gets `426` with `code: client_outdated`.
    - Files are never referenced by a client-supplied URL: upload with `POST /attachments`,
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /attachments:
    parameters:
      - $ref: "#/components/parameters/ClientVersion"
      - $ref: "#/components/parameters/ClientApiVersion"
    post:
      tags: [attachments]
      operationId: uploadAttachment
      summary: Upload a file, image, audio or voice message
      description: |
        The server checks the content, not the file name or the client's content type. The
        returned `attachment_id` is used in messages (`type: file` or `voice`) or for
        avatars. Attachments that are not used within 24 hours are deleted.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: { $ref: "#/components/schemas/UploadAttachmentRequest" }
      responses:
        "426": { $ref: "#/components/responses/ClientOutdated" }
        "201":
          description: Uploaded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Attachment" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
...
  /chats:
    parameters:
      - $ref: "#/components/parameters/ClientVersion"
      - $ref: "#/components/parameters/ClientApiVersion"
    get:
      tags: [chats]
      operationId: listChats
      summary: My chats with last message and unread count
      description: |
        One request replaces `GET /chats/list/{username}` and `GET /groups/list/{username}`.
        Pinned chats come first (most recently pinned first), then the others by their last
        message (newest first). `last_message` is the newest message the caller can see; in
        direct chats its `read_by` holds the other user once they read it (with read receipts
        on), in groups it is empty. `unread_count` stops at 999.
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/After"
      responses:
        "426": { $ref: "#/components/responses/ClientOutdated" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "200":
          description: A page of chats
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChatPage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [chats]
      operationId: createChat
      summary: Start a direct chat
      description: |
        Respects the other user's `direct_messages` privacy setting. When approval is needed
        the answer is `202` with the pending request instead of a chat. Returns the existing
        chat (`200`) if there already is one.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateChatRequest" }
      responses:
        "426": { $ref: "#/components/responses/ClientOutdated" }
        "200":
          description: The chat already existed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Chat" }
        "201":
          description: Chat created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Chat" }
        "202":
          description: Waiting for the other user's approval
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApprovalRequest" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /chats/{chat_id}:
    parameters:
      - $ref: "#/components/parameters/ClientVersion"
      - $ref: "#/components/parameters/ClientApiVersion"
      - $ref: "#/components/parameters/ChatId"
    get:
      tags: [chats]
      operationId: getChat
      summary: One chat
      responses:
        "426": { $ref: "#/components/responses/ClientOutdated" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "200":
          description: The chat
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Chat" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [chats]
      operationId: deleteChat
      summary: Delete a direct chat
      description: |
        Deletes the chat and its messages for both users (as v1 did); both get `chat_deleted`.
        For groups use `DELETE /groups/{chat_id}` (owner) or `POST /groups/{chat_id}/leave`;
        a group here is `422`.
      responses:
        "426": { $ref: "#/components/responses/ClientOutdated" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "204":
          description: Deleted
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /chats/{chat_id}/pin:
    parameters:
      - $ref: "#/components/parameters/ClientVersion"
      - $ref: "#/components/parameters/ClientApiVersion"
      - $ref: "#/components/parameters/ChatId"
    put:
      tags: [chats]
      operationId: pinChat
      summary: Pin a chat
      description: At most 10 chats can be pinned (`409` beyond that). Pinning a pinned chat changes nothing.
      responses:
        "426": { $ref: "#/components/responses/ClientOutdated" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "204":
          description: Pinned
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
    delete:
      tags: [chats]
      operationId: unpinChat
      summary: Unpin a chat
      responses:
        "426": { $ref: "#/components/responses/ClientOutdated" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "204":
          description: Unpinned
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /chats/{chat_id}/read:
    parameters:
      - $ref: "#/components/parameters/ClientVersion"
      - $ref: "#/components/parameters/ClientApiVersion"
      - $ref: "#/components/parameters/ChatId"
    post:
      tags: [chats]
      operationId: markChatRead
      summary: Mark messages as read up to a message
      description: Replaces the duplicate `/read` and `/read/` routes of v1.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MarkReadRequest" }
      responses:
        "426": { $ref: "#/components/responses/ClientOutdated" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "200":
          description: New unread count
          content:
            application/json:
              schema:
                type: object
                required: [unread_count]
                properties:
                  unread_count: { type: integer, minimum: 0 }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /requests:
    parameters:
      - $ref: "#/components/parameters/ClientVersion"
      - $ref: "#/components/parameters/ClientApiVersion"
    get:
      tags: [requests]
      operationId: listApprovalRequests
      summary: Inbox of approval requests
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/After"
      responses:
        "426": { $ref: "#/components/responses/ClientOutdated" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "200":
          description: A page of pending requests
          content:
            application/json:
              schema:
                type: object
                required: [items, next_cursor]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/ApprovalRequest" }
                  next_cursor: { $ref: "#/components/schemas/Cursor" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /requests/{request_id}/approve:
    parameters:
      - $ref: "#/components/parameters/ClientVersion"
      - $ref: "#/components/parameters/ClientApiVersion"
      - $ref: "#/components/parameters/RequestId"
    post:
      tags: [requests]
      operationId: approveRequest
      summary: Approve a request
      responses:
        "426": { $ref: "#/components/responses/ClientOutdated" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "200":
          description: Approved; the chat or group membership now exists
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Chat" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /requests/{request_id}/reject:
    parameters:
      - $ref: "#/components/parameters/ClientVersion"
      - $ref: "#/components/parameters/ClientApiVersion"
      - $ref: "#/components/parameters/RequestId"
    post:
      tags: [requests]
      operationId: rejectRequest
      summary: Reject a request
      responses:
        "426": { $ref: "#/components/responses/ClientOutdated" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "204":
          description: Rejected
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }

  # -------------------------------------------------------------- groups
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived access token from login or refresh.
    cookieAuth:
      type: apiKey
      in: cookie
      name: refresh_token
      description: HttpOnly refresh cookie set by login; scoped to `/api/v2/auth`.
...
  parameters:
    Limit:
      name: limit
      in: query
      description: Page size
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
    After:
      name: after
      in: query
      description: Opaque `next_cursor` of the previous page
      schema: { $ref: "#/components/schemas/Cursor" }
    ChatId:
      name: chat_id
      in: path
      required: true
      schema: { $ref: "#/components/schemas/Id" }
...
    RequestId:
      name: request_id
      in: path
      required: true
      schema: { $ref: "#/components/schemas/Id" }
...
    ClientVersion:
      name: X-Client-Version
      in: header
      required: false
      description: Version of the client app, for logs and the per-client metric.
      schema: { type: string }
    ClientApiVersion:
      name: X-Client-Api-Version
      in: header
      required: false
      description: Contract version the client was built with (`API_VERSION` of `@myronsi/messenger-api`). A different MAJOR or a version below `min_client_api_version` is answered with `426`; a malformed value with `400` (`invalid_client_version`).
      schema: { type: string }
  responses:
    ClientOutdated:
      description: "The `X-Client-Api-Version` header has another MAJOR version than the server or is below `min_client_api_version` (`client_outdated`). Not returned by `/meta`, which clients use to find out the current versions."
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    BadRequest:
      description: Malformed request, or a malformed `X-Client-Api-Version` header (`invalid_client_version`)
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Unauthorized:
      description: Missing, invalid or expired credentials
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Forbidden:
      description: Authenticated but not allowed
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: Does not exist or is not visible to the current user
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Conflict:
      description: The state of the resource does not allow this
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    PayloadTooLarge:
      description: Upload exceeds the size limit
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    UnsupportedMediaType:
      description: The content type is not allowed
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    ValidationFailed:
      description: Syntactically valid but rejected values; `errors` lists the fields
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        Retry-After:
          description: Seconds until the next attempt is allowed
          schema: { type: integer, minimum: 0 }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    Id:
      type: string
      pattern: "^[0-9]{1,19}$"
      description: Decimal ID as a string (Snowflake IDs exceed 2^53).
      examples: ["7217400317439950848"]
    AttachmentRef:
      type: string
      format: uuid
      description: ID of an uploaded attachment (a UUID, unlike the decimal IDs of messages, users and chats).
      examples: ["5f0c8a64-2f2b-4c7e-9d0e-6b1f3a2c4d5e"]
    Cursor:
      type: [string, "null"]
      maxLength: 256
      description: Opaque cursor; `null` when there are no more results.
    Timestamp:
      type: string
      format: date-time
    Username:
      type: string
      pattern: "^[A-Za-z0-9_]{3,32}$"
...
    ErrorCode:
      type: string
      description: |
        Stable machine-readable error code. New codes can appear in MINOR versions; unknown
        codes must be treated like `internal_error`.
      enum:
        - invalid_request
        - validation_failed
        - unauthenticated
        - invalid_credentials
        - two_factor_required
        - invalid_two_factor_code
        - token_expired
        - forbidden
        - not_found
        - conflict
        - already_exists
        - payload_too_large
        - unsupported_media_type
        - rate_limited
        - client_outdated
        - invalid_client_version
        - blocked_by_user
        - approval_required
        - internal_error
    Problem:
      type: object
      description: RFC 9457 problem details. `426 Upgrade Required` (`client_outdated`) uses the same format.
      required: [type, title, status, code]
      properties:
        type:
          type: string
          description: URI reference identifying the problem type (`about:blank` when `code` says it all).
          default: about:blank
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance: { type: string }
        code: { $ref: "#/components/schemas/ErrorCode" }
        errors:
          type: array
          description: Field errors for `validation_failed`.
          items:
            type: object
            required: [field, message]
            properties:
              field:
                type: string
                description: JSON pointer into the request body, or the parameter name.
              message: { type: string }
        min_client_api_version:
          type: string
          description: Present when `code` is `client_outdated`.
        api_version:
          type: string
          description: Present when `code` is `client_outdated`.
      additionalProperties: true
...
    User:
      type: object
      description: A user as visible to the requesting user; hidden fields are `null`.
      required: [id, username, display_name, avatar_url, bio, is_online, last_seen, is_deleted]
      properties:
        id: { $ref: "#/components/schemas/Id" }
        username: { $ref: "#/components/schemas/Username" }
        display_name: { type: string }
        contact_name:
          type: [string, "null"]
          description: The requesting user's own name for this user, if set.
        avatar_url:
          type: [string, "null"]
          description: Path of the avatar, always same-origin (`/api/v2/users/{id}/avatar`).
        bio: { type: [string, "null"] }
        is_online: { type: boolean }
        last_seen:
          oneOf:
            - $ref: "#/components/schemas/Timestamp"
            - type: "null"
        is_deleted: { type: boolean }
...
    AttachmentKind:
      type: string
      enum: [image, audio, voice, video, file]
    UploadPurpose:
      type: string
      enum: [message, avatar]
    UploadAttachmentRequest:
      type: object
      required: [file, purpose]
      properties:
        file:
          type: string
          format: binary
        purpose: { $ref: "#/components/schemas/UploadPurpose" }
        kind: { $ref: "#/components/schemas/AttachmentKind" }
        duration_ms:
          type: integer
          minimum: 0
          description: Length of a voice message
        waveform:
          type: array
          description: Amplitude samples (0-255) of a voice message
          maxItems: 128
          items: { type: integer, minimum: 0, maximum: 255 }
    Attachment:
      type: object
      required: [id, kind, filename, content_type, size, url]
      properties:
        id: { $ref: "#/components/schemas/AttachmentRef" }
        kind: { $ref: "#/components/schemas/AttachmentKind" }
        filename: { type: string }
        content_type:
          type: string
          description: Detected by the server
        size:
          type: integer
          minimum: 0
          description: Bytes
        url:
          type: string
          description: Same-origin path (`/api/v2/attachments/{id}/content`) set by the server.
          readOnly: true
        width: { type: [integer, "null"], minimum: 1 }
        height: { type: [integer, "null"], minimum: 1 }
        thumbnail_url: { type: [string, "null"], readOnly: true }
        duration_ms: { type: [integer, "null"], minimum: 0 }
        waveform:
          type: [array, "null"]
          items: { type: integer, minimum: 0, maximum: 255 }
...
    MessageType:
      type: string
      enum: [text, file, voice, system]
...
    Message:
      type: object
      required:
        - id
        - chat_id
        - type
        - sender
        - content
        - attachment
        - reply_to
        - forwarded_from
        - reactions
        - read_by
        - created_at
        - edited_at
        - is_deleted
      properties:
        id: { $ref: "#/components/schemas/Id" }
        chat_id: { $ref: "#/components/schemas/Id" }
        type: { $ref: "#/components/schemas/MessageType" }
        sender:
          description: "`null` for system messages."
          oneOf:
            - $ref: "#/components/schemas/User"
            - type: "null"
        content:
          type: [string, "null"]
          maxLength: 4096
          description: Text, or the caption of a file; `null` when deleted.
        attachment:
          oneOf:
            - $ref: "#/components/schemas/Attachment"
            - type: "null"
        reply_to:
          oneOf:
            - $ref: "#/components/schemas/Id"
            - type: "null"
        forwarded_from:
          oneOf:
            - $ref: "#/components/schemas/ForwardedFrom"
            - type: "null"
        reactions:
          type: array
          items: { $ref: "#/components/schemas/Reaction" }
        read_by:
          type: array
          items: { $ref: "#/components/schemas/ReadReceipt" }
        created_at: { $ref: "#/components/schemas/Timestamp" }
        edited_at:
          oneOf:
            - $ref: "#/components/schemas/Timestamp"
            - type: "null"
        is_deleted: { type: boolean }
        client_temp_id: { $ref: "#/components/schemas/ClientTempId" }
...
    MarkReadRequest:
      type: object
      required: [message_id]
      properties:
        message_id:
          description: Everything up to and including this message is read.
          $ref: "#/components/schemas/Id"

    # ------------------------------------------------- chats and groups
    ChatType:
      type: string
      enum: [direct, group]
    GroupRole:
      type: string
      enum: [owner, admin, member]
    Chat:
      type: object
      required:
        - id
        - type
        - name
        - avatar_url
        - peer
        - is_pinned
        - unread_count
        - last_message
        - created_at
      properties:
        id: { $ref: "#/components/schemas/Id" }
        type: { $ref: "#/components/schemas/ChatType" }
        name:
          type: string
          description: Display name of the peer or the group
        avatar_url: { type: [string, "null"] }
        peer:
          description: The other user of a direct chat; `null` for groups.
          oneOf:
            - $ref: "#/components/schemas/User"
            - type: "null"
        is_pinned: { type: boolean }
        unread_count: { type: integer, minimum: 0 }
        last_message:
          oneOf:
            - $ref: "#/components/schemas/Message"
            - type: "null"
        my_role:
          description: Only for groups.
          oneOf:
            - $ref: "#/components/schemas/GroupRole"
            - type: "null"
        created_at: { $ref: "#/components/schemas/Timestamp" }
    ChatPage:
      type: object
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Chat" }
        next_cursor: { $ref: "#/components/schemas/Cursor" }
    CreateChatRequest:
      type: object
      required: [user_id]
      properties:
        user_id: { $ref: "#/components/schemas/Id" }
        initial_message:
          type: [string, "null"]
          maxLength: 4096
...
    ApprovalRequestType:
      type: string
      enum: [direct_message, group_invite]
    ApprovalRequestStatus:
      type: string
      enum: [pending, approved, rejected]
    ApprovalRequest:
      type: object
      required: [id, type, status, requester, created_at]
      properties:
        id: { $ref: "#/components/schemas/Id" }
        type: { $ref: "#/components/schemas/ApprovalRequestType" }
        status: { $ref: "#/components/schemas/ApprovalRequestStatus" }
        requester: { $ref: "#/components/schemas/User" }
        group_name:
          type: [string, "null"]
          description: Only for `group_invite`
        preview:
          type: [string, "null"]
          description: First message of a direct-message request
        created_at: { $ref: "#/components/schemas/Timestamp" }